add repository for auditEvent
